* Fixed a possible problem related with write operations in datagram transport. * Fixed a problem that throughput is not updated.
git-svn-id: https://svn.apache.org/repos/asf/mina/trunk@598935 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/mina/transport/socket/nio/NioProcessor.java
@@ -172,8 +172,8 @@ public final class NioProcessor extends AbstractPollingIoProcessor<NioSession> {
     }
 
     @Override
-    protected long transferFile(NioSession session, FileRegion region, int length) throws Exception {
-        return region.getFileChannel().transferTo(region.getPosition(), region.getCount(), session.getChannel());
+    protected int transferFile(NioSession session, FileRegion region, int length) throws Exception {
+        return (int) region.getFileChannel().transferTo(region.getPosition(), region.getCount(), session.getChannel());
     }
 
     protected static class IoSessionIterator implements Iterator<NioSession> {
